Share Vulkan context struct between Win and Skia
