+++ IGraphics/Platforms/IGraphicsWin.h
@@ -25,9 +25,7 @@
 #include <vector>
 
 #ifdef IGRAPHICS_VULKAN
-  #define VK_USE_PLATFORM_WIN32_KHR
-  #include <vulkan/vulkan.h>
-  #include <vulkan/vulkan_win32.h>
+  #include "VulkanContext.h"
   #include "WinVulkanDeviceCoordinator.h"
 #endif
 
@@ -36,25 +34,7 @@ BEGIN_IPLUG_NAMESPACE
 BEGIN_IGRAPHICS_NAMESPACE
 
 #ifdef IGRAPHICS_VULKAN
-struct VulkanContext
-{
-  VkInstance instance = VK_NULL_HANDLE;
-  VkPhysicalDevice physicalDevice = VK_NULL_HANDLE;
-  VkDevice device = VK_NULL_HANDLE;
-  VkSurfaceKHR surface = VK_NULL_HANDLE;
-  VkSwapchainKHR swapchain = VK_NULL_HANDLE;
-  VkQueue queue = VK_NULL_HANDLE;
-  uint32_t queueFamily = 0;
-  VkSemaphore imageAvailableSemaphore = VK_NULL_HANDLE;
-  VkSemaphore renderFinishedSemaphore = VK_NULL_HANDLE;
-  VkFence inFlightFence = VK_NULL_HANDLE;
-  std::vector<VkImage>* swapchainImages = nullptr;
-  VkFormat format = VK_FORMAT_B8G8R8A8_UNORM;
-  VkImageUsageFlags usageFlags = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-#if IGRAPHICS_SANDBOX_LOGGING
-  const vulkanlog::LoggerContext* loggerContext = nullptr;
-#endif
-};
+using VulkanContext = ::iplug::igraphics::VulkanContext;
 #endif
 
 // Forward declare the OLE drop target helper (defined in IGraphicsWin_dnd.h)
