+++ Plan/Plan-Summary.md
@@ -1,19 +1,16 @@
-[x] Continued From Previous Snapshot: YES — Continued PLAN-EXECUTION by sandboxing Skia Vulkan context storage, enforcing draw umbrella cache isolation, and refreshing the Windows sandbox documentation.
+[x] Continued From Previous Snapshot: YES — Updated PLAN-EXECUTION scope to include repairing Skia/Vulkan Windows compilation failures introduced by ctx/VulkanContext regressions.
 -----------------
 [x] File Overview:
-- Num files Changed: 7
+- Num files Changed: 4
 - Num files Created: 1
-- Num lines Modified: 353
+- Num lines Modified: 111
 
 [x] List of Files changed/created:
-- Documentation/WindowsSandbox.md
-- IGraphics/Drawing/IGraphicsSkia.cpp
-- IGraphics/Drawing/IGraphicsSkia.h
-- IGraphics/IGraphics.cpp
-- IGraphics/IGraphics.h
 - Plan/Current-Plan.xml
 - Plan/Plan-Summary.md
-- Plan/State-of-Sandbox-Report.md
+- IGraphics/Drawing/IGraphicsSkia.cpp
+- IGraphics/Platforms/IGraphicsWin.h
+- IGraphics/Platforms/VulkanContext.h
 
 -----------------
 [x] Current plan:
@@ -28,7 +25,8 @@
 - IGRAPHICS_SANDBOX_LOGGING — PREVIOUS STATUS: OPEN / CURRENT STATUS: SUCCESS
 - IGRAPHICS_SANDBOX_VK_LOGGER — PREVIOUS STATUS: OPEN / CURRENT STATUS: SUCCESS
 - IGRAPHICS_SANDBOX_VK_LOG_LEVEL — PREVIOUS STATUS: OPEN / CURRENT STATUS: SUCCESS
+- FIX_VULKAN_CTX_COMPILATION_ERRORS — PREVIOUS STATUS: N/A / CURRENT STATUS: AWAIT-CHILDREN-TASK-SUCCESS
 - FINAL CHECK — PREVIOUS STATUS: OPEN / CURRENT STATUS: OPEN
 
 [x] Message to User:
-Sandbox builds now route Skia's Vulkan command pools, swapchain images, and semaphores through per-instance storage while allowing legacy runs to keep the shared singleton when the toggle is disabled. The draw umbrella switch automatically promotes bitmap/SVG caches to renderer-owned storage, and the Windows sandbox guide documents the new expectations for hosts awaiting Windows validation.
+Shared the VulkanContext struct between the Windows platform layer and Skia renderer so ctx resolves during compilation; next step is to document the Windows validation path for the outstanding verification subtask.
